refactor(api): move classes to the correct package
