--- api/src/main/java/com/znaji/event_ticket_platform/domain/OrderStatus.java
@@ -1,4 +1,0 @@
-package com.znaji.event_ticket_platform.domain;
-
-public enum OrderStatus {
-}
