+++ api/src/main/java/com/znaji/event_ticket_platform/domain/events/TicketType.java
@@ -1,4 +1,4 @@
-package com.znaji.event_ticket_platform.domain;
+package com.znaji.event_ticket_platform.domain.events;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
