+++ api/src/main/java/com/znaji/event_ticket_platform/domain/orders/OrderStatus.java
@@ -1,0 +1,4 @@
+package com.znaji.event_ticket_platform.domain.orders;
+
+public enum OrderStatus {
+}
